requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
pymongo>=4.0.0
flask-pymongo>=2.3.0
python-dotenv>=0.19.0
bcrypt>=3.2.0
flask-jwt-extended>=4.4.0
